refactor: split home screen sections
