+++ src/features/home/screens/HomeProducts.styles.ts
@@ -1,0 +1,42 @@
+import { StyleSheet } from 'react-native';
+import { spacing } from '@/shared/ui/tokens';
+
+export const styles = StyleSheet.create({
+  section: {
+    paddingHorizontal: spacing.spacer16,
+    marginBottom: spacing.spacer24,
+  },
+  sectionHeader: {
+    justifyContent: 'space-between',
+    alignItems: 'center',
+    marginBottom: spacing.spacer16,
+  },
+  sectionTitle: {
+    fontSize: 18,
+    fontWeight: 'bold',
+  },
+  sectionActions: {
+    alignItems: 'center',
+  },
+  sortButton: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    paddingHorizontal: spacing.spacer12,
+    paddingVertical: 6,
+    borderRadius: 16,
+    borderWidth: 1,
+  },
+  sortText: {
+    fontSize: 12,
+    fontWeight: '500',
+    marginLeft: spacing.spacer4,
+  },
+  addProductButton: {
+    width: 28,
+    height: 28,
+    borderRadius: 14,
+    justifyContent: 'center',
+    alignItems: 'center',
+    borderWidth: 1,
+  },
+});
